requirements.txt
# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
sse-starlette>=1.6.0

# Document parsing
pdfplumber>=0.10.0
python-docx>=1.0.0

# Database
aiosqlite>=0.19.0

# Embedding & search (Phase D2)
sentence-transformers>=2.2.0
qdrant-client>=1.7.0
rank-bm25>=0.2.2
kiwipiepy>=0.16.0
numpy>=1.26.0

# LLM client (Phase D3)
openai>=1.0.0

# LLM server (Phase D3 — serve_llm.py)
llama-cpp-python>=0.3.0
